Split the layouts to fix height and scrolling
