+++ src/pages/image/_id/index.vue
@@ -137,6 +137,7 @@ export default defineComponent({
     }
     next()
   },
+  layout: "content-layout",
   setup() {
     const route = useRoute()
