+++ src/pages/search-help.vue
@@ -236,6 +236,7 @@ import VContentPage from "~/components/VContentPage.vue"
 export default defineComponent({
   name: "VSearchHelpPage",
   components: { VLink, VContentPage },
+  layout: "content-layout",
   setup() {
     const i18n = useI18n()
     const featureFlagStore = useFeatureFlagStore()
